# docs/api-openapi.yaml
openapi: 3.0.3
info:
  title: x-bot API
  description: X (Twitter) automation platform API
  version: 0.1.0
servers:
  - url: http://localhost:3000/api
paths:
  /auth/login:
    post:
      summary: Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
...
  /accounts/{id}/actions:
    post:
      summary: Execute action on account
      tags: [Accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [warmup, like, retweet, comment]
      responses:
        '200':
          description: Action executed
  /campaigns:
    get:
      summary: List campaigns
      tags: [Campaigns]
      responses:
        '200':
          description: List of campaigns
    post:
      summary: Create campaign
      tags: [Campaigns]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [urls, comments, browsersCount, accountIds]
              properties:
                urls:
                  type: array
                  items:
                    type: string
                comments:
                  type: array
                  items:
                    type: string
                browsersCount:
                  type: integer
                accountIds:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Campaign created
...
  /reports:
    get:
      summary: Export report
      tags: [Reports]
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [campaigns, accounts]
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv]
      responses:
        '200':
          description: Report data
...